fix: separate types into folders are export types for imports
