+++ src/native/pure.ts
@@ -1,10 +1,10 @@
 import { act, create, ReactTestRenderer } from 'react-test-renderer'
 
-import { RendererProps } from 'types'
-import { ReactRendererOptions } from '../react/types'
+import { RendererProps } from '../types'
+import { ReactRendererOptions } from '../types/react'
 
 import { createRenderHook, cleanup, addCleanup, removeCleanup } from '../core'
-import { createTestHarness } from '../react/createTestHarness'
+import { createTestHarness } from '../helpers/createTestHarness'
 
 function createNativeRenderer<TProps, TResult>(
   testHookProps: RendererProps<TProps, TResult>,
@@ -37,3 +37,6 @@ function createNativeRenderer<TProps, TResult>(
 const renderHook = createRenderHook(createNativeRenderer)
 
 export { renderHook, act, cleanup, addCleanup, removeCleanup }
+
+export * from '../types'
+export * from '../types/react'
